daily-agent/requirements.txt
langchain
langchain-community
openai
serpapi
python-dotenv
transformers
torch
